Avoid admin API entirely
Calls to admin_externalpage_setup() when logged in as a user without the
moodle:site/config capability result in section errors, as Moodle
doesn't initialise the settings_navigation tree in its entirety. Since a
fundamental requirement of the licensing plugin is that the allocation
interface should be accessible to non-administrators, we'll move to
manually checking our capabilities instead.

This change also introduces checking of capabilities in edit.php.